+++ box-ui/src/components/css/CaptureImage.css
