+++ box-ui/src/components/ts/CaptureImage.tsx
